Released the Quiet HN Exercise
This commit contains the README updates necessary to describe
the Quiet HN exercise, as well as the code required for the
initial state of the exercise. This includes a simple Hacker
News API client, tests and examples for the client, and a
working server (see main.go) that implements the exercise
WITHOUT caching or concurrency.